requirements.txt
aiogram
sqlalchemy
python-dotenv
asyncpg
aiosqlite


# pip install -r requirements.txt
